dir_iterator over a symlink should always follow the symlink.
It is not crystal clear in the C++ standard what is the exact interpretation
of the follow_directory_symlink in directory_options. Our interpretation is
that it is not significant for dir_iterator and it is only used for
dir_recursive_iterator. This is in line with the default behavior of 'ls'
command and with the fact that POSIX opendir always follows symlinks in the
given path.